Equilibrado y build settings

+++ Assets/Scripts/Wife.cs
@@ -9,7 +9,7 @@ public class Wife : MonoBehaviour {
 	public Slider affinityBar;
 	[SerializeField] private int maxAffinity;
 	private int startValue;
-	private int affinity;
+	private float affinity;
 
 
 
@@ -23,7 +23,7 @@ public class Wife : MonoBehaviour {
 	// ------ Getters & Setters
 
 
-	public void setAffinity(int nAffinity){
+	public void setAffinity(float nAffinity){
 
 		affinity = nAffinity;
 		affinityBar.maxValue = maxAffinity;
@@ -31,13 +31,13 @@ public class Wife : MonoBehaviour {
 		affinityBar.value = startValue;
 	}
 
-	public int getAffinity(){
+	public float getAffinity(){
 
 		return affinity;
 
 	}
 
-	public void UpdateAffinity(int nAffinity){
+	public void UpdateAffinity(float nAffinity){
 	
 		//affintyBar.value += Mathf.MoveTowards(0, nAffinity, 0.1f);
 		if(affinityBar.value + nAffinity > maxAffinity){
